Пока 5.0.0.7 проходит  BigCrush нет предела совершенству. 5.0.0.8

+++ perfect_random.hpp
@@ -223,7 +223,7 @@ public:
 
     // ── MeanCore::adjust — исправленная нормализация ──
     // MeanCore::adjust — ЗАМЕНИТЬ:
-    double adjust(unsigned int base) {
+    double adjust(double base) {
         return (base / static_cast<double>(UINT_MAX)) * mean + rng.generate();
     }
